+++ src/main/java/com/bayudwiyansatria/plot/PlotPanel.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package com.bayudwiyansatria.utils;
+package com.bayudwiyansatria.plot;
 import java.awt.Color;
 import java.awt.Font;
 import java.awt.Graphics;
